+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 
 - Accept [special initial conversation `current`](https://github.com/speee/jsx-slack/blob/master/docs/block-elements.md#special-initial-conversation-current) in `<ConversationsSelect>` ([#154](https://github.com/speee/jsx-slack/issues/154), [#155](https://github.com/speee/jsx-slack/pull/155))
 
+### Changed
+
+- Upgrade Node and dependent packages to the latest version ([#158](https://github.com/speee/jsx-slack/pull/158))
+- Update demo REPL ([#157](https://github.com/speee/jsx-slack/pull/157))
+  - Move template examples on README into REPL demo
+  - Use Web fonts to get better rendering
+  - Disable preview button if Slack may return 414 error due to too long URL
+
 ## v2.0.0 - 2020-04-23
 
 <p align="center">